requirements.txt
apache-airflow
black
docker
pylint
